+++ RELEASE_NOTES.md
@@ -5,3 +5,7 @@
 
 ## 0.1.0-SNAPSHOT
 * Added SpringBoot skeleton project
+* Added telegram bot to the project
+
+## 0.2.0-SNAPSHOT
+* Added Command pattern